@@ -2,16 +2,16 @@ package fr.fullstack.shopapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.domain.Pageable;
-import org.springframework.validation.Errors;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.data.domain.Pageable;
+//import org.springframework.validation.Errors;
+//import springfox.documentation.builders.PathSelectors;
+//import springfox.documentation.builders.RequestHandlerSelectors;
+//import springfox.documentation.spi.DocumentationType;
+//import springfox.documentation.spring.web.plugins.Docket;
+//import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@EnableSwagger2
+//@EnableSwagger2
 @SpringBootApplication
 public class ShopAppApplication {
 
@@ -19,14 +19,15 @@ public class ShopAppApplication {
         SpringApplication.run(ShopAppApplication.class, args);
     }
 
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .ignoredParameterTypes(Errors.class, Pageable.class)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("fr.fullstack.shopapp.controller"))
-                .paths(PathSelectors.any())
-                .build();
-    }
+//    @Bean
+//    public Docket api() {
+//        return new Docket(DocumentationType.SWAGGER_2)
+//                .ignoredParameterTypes(Errors.class, Pageable.class)
+//                .select()
+//                .apis(RequestHandlerSelectors.basePackage("fr.fullstack.shopapp.controller"))
+//                .paths(PathSelectors.any())
+//                .build();
+//    }
+
 
 }
